refactor: adjust compat code
