+++ editors/vscode/src/extension.preview.ts
@@ -1,12 +1,12 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
-import { previewActive, previewDeactivate } from "./preview";
+import { previewActivate, previewDeactivate } from "./preview";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-    return previewActive(context, true);
+    return previewActivate(context, true);
 }
 
 // This method is called when your extension is deactivated
